Mehr Verknüpfungen zwischen Controller und xml
